backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
openpyxl==3.1.2
reportlab==4.0.8
python-dotenv==1.0.0
Werkzeug==3.0.1
pandas==2.3.3
